[Refactor] split helm code across a few files for readability
